remove enums
